+++ src/cpu-pricer.cpp
@@ -6,11 +6,11 @@
 int main(int argc, char *argv[]) {
     if (argc == 7) {
         // Parse arguments
-        double X0 = std::atof(argv[1]);
-        double K = std::atof(argv[2]);
-        double r = std::atof(argv[3]);
-        double sigma = std::atof(argv[4]);
-        double T = std::atof(argv[5]);
+        float X0 = std::atof(argv[1]);
+        float K = std::atof(argv[2]);
+        float r = std::atof(argv[3]);
+        float sigma = std::atof(argv[4]);
+        float T = std::atof(argv[5]);
         int N = std::atoi(argv[6]);
 
         // Create american option
@@ -21,9 +21,9 @@ int main(int argc, char *argv[]) {
 
         // Pricing with timer
         std::clock_t start = std::clock();
-        double price = option->pricing();
+        float price = option->pricing();
         std::clock_t end = std::clock();
-        double duration = (double)(end - start) / CLOCKS_PER_SEC;
+        float duration = (float)(end - start) / CLOCKS_PER_SEC;
 
         // Print price at 0
         std::cout << "Price: " << price << std::endl;
